move stake records into sub_files, code clean up
